Implement support for handling large decimal values in database and API requests
Changes include:
- Updated the database schema by changing the 'updated_value' column from INT to DECIMAL using Alembic migrations to accommodate larger numerical values.
- Added a custom JSON encoder, DecimalEncoder, to serialize Decimal objects into JSON data properly. This encoder is now integrated into the _request method of the Api class, ensuring that Decimal types are serialized as floats during API interactions.
- Adjusted Alembic configurations to autogenerate migrations based on model changes, enhancing our migration management process.

+++ backend/db/models/node_aggregations.py
@@ -1,6 +1,7 @@
 """Model for the node aggregations entity."""
 
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional
 from uuid import UUID
 
@@ -16,7 +17,7 @@ class NodeAggregationBase(SQLModel):
     feed_id: UUID = Field(foreign_key="feed.id")
     timestamp: datetime
     status: str
-    aggregated_value: Optional[int] = Field(default=None, nullable=True)
+    aggregated_value: Optional[Decimal] = Field(default=None, nullable=True)
     nodes_count: Optional[int] = Field(default=None, nullable=True)
     tx_hash: Optional[str] = Field(default=None, nullable=True, index=True)
     trigger: Optional[str] = Field(default=None, nullable=True)
